requirements.txt
# WARNING !!!
# =============================================================================
#
# This file is auto-generated from environment.yml file, do not modify it directly.
#
# See in `environment.yml` for more information.
#
# =============================================================================

quadprog
brukeropusreader
numpy-quaternion
colorama
docrep
dill
ipython>=7.31.1
jinja2<3.1
matplotlib>=3.5
numba==0.55.0
numpy==1.21.5
pint>=0.18
requests
scipy==1.7.3
tqdm
traitlets
traittypes
xlrd
pyyaml
ipywidgets
ipympl
setuptools
setuptools_scm
gitpython
